feat(main): fix  error for not healthy
